+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 4.1.1 - 2020-01-11
+
+**Fixed**
+
+- Fixed inline attachments could not be stored
+- Fixed PHP 7.4 issue when reading empty Mailable from address
+
 ## 4.1.0 - 2019-07-13
 
 **Added**